🛒 +buy max
The buy command now supports arguments max and all to purchase as many of an item as the user can afford. I tested this code and it works, I also made it double check the balance so you can't transfer your money during the confirmation period as is an issue with several other commands.